some ids are longer than 10 digits which cause conversion issues on linux for int32 values. Needed to change some of them to let my app running smoothly. May not be the final fix yet

+++ rest/private/orders/cancel.go
@@ -7,7 +7,7 @@ import (
 
 type RequestForCancelByID struct {
 	ClientID       string
-	OrderID        int
+	OrderID        int64
 	TriggerOrderID string
 }
